requirements.txt
alpha_vantage==3.0.0
pandas==1.4.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
torch==2.2.2
transformers==4.48.3
apache-airflow==2.10.5